+++ src/components/pages/campaign/components/batches-list/index.tsx
@@ -37,7 +37,11 @@ const defineDispenserTypes = (
   batchId: string,
   tokenAddress: string,
   wallet: string,
+  customClaimHost: string,
   campaignTitle: string,
+
+  dispenserMappingPageRedirect: () => void,
+  qrSetMappingPageRedirect: () => void,
 
   successCallbackForDispenser?: (
     dispenser_id: string | number,
@@ -56,12 +60,14 @@ const defineDispenserTypes = (
       text: 'A web page with an auto-refresh QR code that updates in real time. This ensures secure distribution, preventing a single user from claiming all tokens',
       onClick: () => {
         createDispenserAndAddLinks(
+          dispenserMappingPageRedirect,
           `Dispenser for ${campaignTitle}`,
           true,
           campaignId,
           batchId,
           tokenAddress,
           wallet,
+          customClaimHost,
           successCallbackForDispenser,
           errorCallback
         )
@@ -72,12 +78,14 @@ const defineDispenserTypes = (
       text: 'A single QR code that dispenses tokens one-by-one to users after they scan it. Ideal for controlled and sequential token distribution',
       onClick: () => {
         createDispenserAndAddLinks(
+          dispenserMappingPageRedirect,
           `Dispenser for ${campaignTitle}`,
           false,
           campaignId,
           batchId,
           tokenAddress,
           wallet,
+          customClaimHost,
           successCallbackForDispenser,
           errorCallback
         )
@@ -88,11 +96,13 @@ const defineDispenserTypes = (
       text: 'A set of single-claim QR codes. Each QR code is valid for one claim only, and becomes invalid after being scanned and claimed by a user',
       onClick: () => {
         createQRSetAndAddLinks(
+          qrSetMappingPageRedirect,
           `QR set for ${campaignTitle}`,
           campaignId,
           batchId,
           tokenAddress,
           wallet,
+          customClaimHost,
           successCallbackForQRSet
         )
       },
@@ -124,6 +134,8 @@ const BatchesList: FC<TProps> = ({
   sdk,
   loading,
   downloadLinks,
+  customClaimHost,
+  customClaimHostOn,
   tokenAddress,
   encryptionKey,
   sponsored,
@@ -152,29 +164,41 @@ const BatchesList: FC<TProps> = ({
     String(showPopup),
     tokenAddress as string,
     wallet,
+  
+    customClaimHost,
+
     title,
+
+    () => {
+      history.push(`/campaigns/${campaignId}/dispenser/generate`)
+    },
+
+    () => {
+      history.push(`/campaigns/${campaignId}/qrs/generate`)
+    },
 
     // for dispenser
     (
       dispenser_id: string | number,
       dynamic: boolean
     ) => {
-        setShowPopup(false)
-        if (dynamic) {
-          return history.push(`/dynamic-qrs/${dispenser_id}`)
-        }
-        return history.push(`/dispensers/${dispenser_id}`)
-      },
+      setShowPopup(false)
+      if (dynamic) {
+        return history.push(`/dynamic-qrs/${dispenser_id}`)
+      }
+      return history.push(`/dispensers/${dispenser_id}`)
+    },
     // for qr-set
     (
       dispenser_id: string | number
     ) => {
-        setShowPopup(false)
-        return history.push(`/qrs/${dispenser_id}`)
-      },
+      setShowPopup(false)
+      return history.push(`/qrs/${dispenser_id}`)
+    },
     () => {
-        setShowPopup(false)
-      },
+      setShowPopup(false)
+      return history.push(`/campaigns/${campaignId}`)
+    },
   )
 
   return <>
@@ -230,7 +254,7 @@ const BatchesList: FC<TProps> = ({
                 }}
               />
 
-              {false && <WidgetButton
+              <WidgetButton
                 appearance='additional'
                 disabled={batch.claim_links_count === 0}
                 size='extra-small'
@@ -238,7 +262,7 @@ const BatchesList: FC<TProps> = ({
                 onClick={() => {
                   setShowPopup(batch.batch_id)
                 }}
-              />}
+              />
             </Buttons>
             
           </>