+++ src/components/pages/campaign/components/settings/custom-claim-host/types.tsx
@@ -1,25 +1,16 @@
-import { TTokenType } from "types"
-
 export type TProps = {
   title: string
   subtitle: string
   onClose: () => void
   id: string
-  preferredWalletValue: string
-
-  tokenType: TTokenType
-
-  sponsored: boolean
-
-  chainId: number
-
-  loading: boolean
+  customClaimHost: string
 
   action: (
-    walletPreferredValue: string,
+    claimHost: string,
     successAction?: () => void,
     errorAction?: () => void
   ) => void
+
 
   toggleAction?: (value: boolean) => void
   toggleValue?: boolean