+++ src/components/pages/campaigns-create-secure/components/index.tsx
@@ -1,5 +1,0 @@
-import CountriesList from './countries-list'
-
-export {
-  CountriesList
-}