+++ src/data/store/reducers/campaign/async-actions/generate-links.tsx
@@ -67,7 +67,8 @@ const generateERC20Link = ({
         preferredWalletOn,
         collectionId,
         collectionTokenId,
-        countriesWhitelistOn
+        countriesWhitelistOn,
+        additionalWalletsOn
       } = campaign
 
       if (!assets) { return alertError('assets are not provided') }
@@ -176,8 +177,12 @@ const generateERC20Link = ({
           available_countries: countriesWhitelist,
           available_countries_on: countriesWhitelistOn,
           preferred_wallet_on: preferredWalletOn,
+          additional_wallets_on: additionalWalletsOn,
           collection_id: collectionId ? collectionId : undefined,
           collection_token_id: collectionTokenId ? collectionTokenId : undefined,
+          claim_host: '',
+          claim_host_on: false,
+          multiple_claims_on: false,
           ...batch
         }