+++ src/data/store/reducers/campaign/types.tsx
@@ -37,6 +37,11 @@ export interface CampaignState {
   // for own collections
   collectionId?: null | string,
   collectionTokenId?: null | string
+  additionalWalletsOn: boolean
+
+  claimHost: string | null
+  claimHostOn: boolean
+  multipleClaimsOn: boolean
 }
 
 export type CampaignActions = ActionType<typeof actions>;