+++ src/data/store/reducers/campaigns/async-actions/index.tsx
@@ -10,9 +10,17 @@ import updateClaimingFinishedButton from './update-claiming-finished-button'
 import updateAvailableCountries from './update-available-countries'
 import updateWallets from './update-wallets'
 import updatePreferredWalletOn from './update-preferred-wallet-on'
+import updateAdditionalWalletsOn from './update-additional-wallets-on'
+import updateClaimHostOn from './update-claim-host-on'
+import updateClaimHost from './update-claim-host'
+import updateMultipleClaimsOn from './update-multiple-claims-on'
 
 export {
   getCampaignBatches,
+  updateClaimHostOn,
+  updateAdditionalWalletsOn,
+  updateMultipleClaimsOn,
+  updateClaimHost,
   updateWallets,
   downloadLinks,
   updateAvailableCountries,