+++ src/data/store/reducers/user/async-actions/secure.tsx
@@ -24,6 +24,8 @@ const secure = (
   availableСountries: TCountry[],
   availableCountriesOn: boolean,
   expirationDate: number,
+  additionalWalletsOn: boolean,
+
   successCallback?: () => void
 ) => {
 
@@ -136,7 +138,12 @@ const secure = (
           dispatch(campaignActions.setSecured(true))
           dispatch(campaignActions.setCountriesWhitelist(availableСountries.map(country => country.id)))
           dispatch(campaignActions.setCountriesWhitelistOn(availableCountriesOn))
+          dispatch(campaignActions.setAdditionalWalletsOn(additionalWalletsOn))
           dispatch(campaignActions.setExpirationDate(expirationDate))
+
+          dispatch(campaignActions.setAdditionalWalletsOn(additionalWalletsOn))
+          dispatch(campaignActions.setExpirationDate(expirationDate))
+
           dispatch(campaignActions.setNativeTokensPerLink(
             utils.parseEther(
               String(